+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@ Release Notes
 =============
 
 ## 1.9.7
-- AKS Cluster: support for Sku and Tier. Support for pod subnet in agent pool config
+- AKS Cluster: support for Sku and Tier. Support for pod subnet in agent pool config. Support for node pool autoscaling
 
 ## 1.9.6
 - Network Interface: Support for adding Network Security Group (NSG) to Network Interface (NIC)